fixed hash sign with torrent downloads

+++ src/filelister/filelistdialog.cpp
@@ -118,7 +118,7 @@ bool FileListDialog::AddTorrentRow(TorrentTable::PRow data)
       m_filelistctrl->SetItemText( index, data->name );
 
       //setting hash as item's data means we can retrieve it later for download
-      m_filelistctrl->SetItemData( index, s2l(data->hash) );
+      m_filelistctrl->SetItemData( index, (unsigned int)s2l(data->hash) );
       m_filelistctrl->SetItem( index, 0, data->name );
       m_filelistctrl->SetItem( index, 1, data->type == IUnitSync::map ? _("Map") : _("Mod") );
       m_filelistctrl->SetItem( index, 2, data->hash );
